requirement.txt
accelerate==1.10.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioitertools==0.12.0
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
arize-phoenix==11.20.0
arize-phoenix-client==1.14.2
arize-phoenix-evals==0.26.1
arize-phoenix-otel==0.12.1
attrs==25.3.0
audioread==3.0.1
Authlib==1.6.1
av==15.0.0
beautifulsoup4==4.13.4
boto3==1.40.4
botocore==1.40.4
Brotli==1.1.0
cachetools==6.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
cryptography==45.0.6
ctranslate2==4.6.0
ddgs==9.5.2
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
e2b==1.11.1
e2b-code-interpreter==1.5.2
email_validator==2.2.0
fastapi==0.116.1
faster-whisper==1.2.0
ffmpy==0.6.1
filelock==3.13.1
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2024.6.1
googleapis-common-protos==1.70.0
gradio==5.41.1
gradio_client==1.11.0
graphql-core==3.2.6
greenlet==3.2.4
groovy==0.1.2
grpc-interceptor==0.15.4
grpcio==1.74.0
h11==0.16.0
helium==5.1.1
hf-xet==1.1.7
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.3
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.4
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonref==1.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
lazy_loader==0.4
librosa==0.11.0
litellm==1.75.0
llvmlite==0.44.0
lxml==6.0.0
Mako==1.3.10
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==2.1.5
mcp==1.12.3
mcpadapt==0.1.12
mdurl==0.1.2
mlx==0.28.0
mlx-lm==0.26.3
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
networkx==3.3
numba==0.61.2
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
onnxruntime==1.22.1
openai==1.99.1
openinference-instrumentation==0.1.37
openinference-instrumentation-smolagents==0.1.14
openinference-semantic-conventions==0.1.21
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.1
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
pillow==11.0.0
piper-tts==1.3.0
platformdirs==4.3.8
pooch==1.8.2
primp==0.15.0
prometheus_client==0.22.1
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pvporcupine==3.0.5
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
PySocks==1.7.1
pystache==0.6.8
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
rich==14.1.0
rpds-py==0.27.0
ruff==0.12.7
s3transfer==0.13.1
safehttpx==0.1.6
safetensors==0.6.1
scikit-learn==1.7.1
scipy==1.16.1
selenium==4.34.2
semantic-version==2.10.0
setuptools==70.2.0
shellingham==1.5.4
six==1.17.0
smolagents==1.21.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.2
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
SQLAlchemy==2.0.42
sqlean.py==3.49.1
sse-starlette==3.0.2
starlette==0.47.2
strawberry-graphql==0.270.1
sympy==1.13.1
threadpoolctl==3.6.0
tiktoken==0.10.0
tokenizers==0.21.4
tomlkit==0.13.3
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.55.0
trio==0.30.0
trio-websocket==0.12.2
triton==3.1.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.1
zipp==3.23.0
